Added AQI to datastore.
When PM2.5 data is imported, the AQI is calculated and added as well.

Air quality getter functions now return the calculated AQI as well. This
required some refactoring of some utility functions:
- Added getFullIDFromRedisKey: this returns the primary ID as well the
full key path.
- Removed getTimeIndexFromSortedSet and replaced it with
marshalRawDataFromResults. This new function now scans the entire
result set from a pipeline and progressively constructs the response
structs. This new method should be roughly the same speed, but consumes
more memory to create lookup tables so that correct result set can be
created.

+++ internal/purpleapi/purple_api.go
@@ -58,7 +58,7 @@ type Response struct {
 	LastUpdated int64    `json:"LastSeen" db:"value,quality"`
 	Latitude    float64  `json:"Lat" db:"value,geo"`
 	Longitude   float64  `json:"Lon" db:"value,geo"`
-	PM25        float32  `json:"PM2_5Value,string" db:"value,quality"`
+	PM25        float64  `json:"PM2_5Value,string" db:"value,quality"`
 }
 
 // UnmarshalJSON implements a custom unmarshaller for the Response type.
